requirements.txt
# Facet Dependencies
# Install with: pip install -r requirements.txt
# Note: Also requires 'exiftool' command-line tool (install via system package manager)

# Deep Learning / GPU
# PyTorch is installed separately with the correct CUDA index URL.
# Run: bash install.sh (auto-detects GPU)
# Or manually: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
open-clip-torch>=2.20.0

# HuggingFace Transformers - required for BiRefNet saliency, SigLIP 2 NaFlex embeddings,
# and VLM taggers (Qwen3-VL / Qwen2.5-VL on 16gb/24gb profiles)
transformers>=4.57.0
# accelerate>=0.25.0              # Optional: needed for VLM taggers with device_map="auto"

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
pillow-heif>=0.16.0
imagehash>=4.3.0
rawpy>=0.18.0

# Face Detection
insightface>=0.7.0
# ONNX Runtime - choose ONE based on your setup:
# onnxruntime>=1.15.0           # CPU only (works everywhere)
# onnxruntime-gpu>=1.17.0       # GPU acceleration for CUDA 12.x
# onnxruntime-gpu>=1.15.0,<1.18 # GPU acceleration for CUDA 11.8

# Quality Assessment (PyIQA models: TOPIQ, HyperIQA, DBCNN, MUSIQ)
pyiqa>=0.1.10

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
hdbscan>=0.8.40  # Face clustering (standalone library, sklearn's HDBSCAN has epsilon bug)

# Web Interface
fastapi>=0.100.0
uvicorn>=0.23.0
pyjwt>=2.8.0

# Utilities
tqdm>=4.65.0
exifread>=3.0.0

# System Monitoring (for batch processing auto-tuning)
psutil>=5.9.0

# Reverse Geocoding (offline, local GeoNames dataset ~30MB)
reverse_geocoder>=1.5.1

# Vector search (sqlite-vec for CLIP/SigLIP embedding KNN queries)
sqlite-vec>=0.1.6

# Subject Saliency: BiRefNet loaded via transformers (see above)

# GPU Face Clustering (optional, requires conda + CUDA)
# cuml
# cupy
